Created a commands file to house all commands
